adding in custom rendering and start of theme
