+++ .env
@@ -12,4 +12,4 @@ NEXT_PUBLIC_RPC_URL=https://eth-mainnet.alchemyapi.io/v2/TMmFiIhF3-KTab6spgLh-RR
 
 ## MAINNET ENV VARS
 NEXT_PUBLIC_NETWORK_ID=1
-NEXT_PUBLIC_TARGET_CONTRACT_ADDRESS=0x8707276df042e89669d69a177d3da7dc78bd8723
+NEXT_PUBLIC_TARGET_CONTRACT_ADDRESS=0x8707276DF042E89669d69A177d3DA7dC78bd8723
