+++ components/AuctionsList.tsx
@@ -1,6 +1,5 @@
-import { FetchStaticData } from "@zoralabs/nft-hooks";
-import { NFTPreview } from "@zoralabs/nft-components";
 import { useRouter } from "next/router";
+import { TokenThumbnail } from "./CustomThumbnail";
 
 export const AuctionsList = ({ tokens }: { tokens: any[] }) => {
   const router = useRouter();
@@ -9,21 +8,8 @@ export const AuctionsList = ({ tokens }: { tokens: any[] }) => {
     <div css={{ display: "flex", flexWrap: "wrap", justifyContent: "center" }}>
       {tokens &&
         tokens.map((token) => {
-          const tokenInfo = FetchStaticData.getIndexerServerTokenInfo(token);
-          return (
-            <NFTPreview
-              initialData={token}
-              key={tokenInfo.tokenId}
-              id={tokenInfo.tokenId}
-              contract={tokenInfo.tokenContract}
-              onClick={(evt) =>
-                router.push(
-                  `/token/${tokenInfo.tokenContract}/${tokenInfo.tokenId}`
-                )
-              }
-              useBetaIndexer={true}
-            />
-          );
+          // const tokenInfo = FetchStaticData.getIndexerServerTokenInfo(token);
+          return <TokenThumbnail token={token} />;
         })}
     </div>
   );
