+++ package.json
@@ -15,7 +15,10 @@
     "@zoralabs/nft-hooks": "^0.5.1",
     "@zoralabs/simple-wallet-provider": "^0.0.7",
     "classnames": "^2.3.1",
+    "dope-metrics": "^1.0.2",
+    "gear-rarity": "^1.0.5",
     "next": "10.2.3",
+    "parse-data-url": "^4.0.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "remark-parse": "^9.0.0",
